requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dateutil==2.8.2
pytz==2023.3
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
certifi==2023.7.22
requests==2.31.0
geoip2==4.7.0
maxminddb==2.2.0
